+++ src/profiler.h
@@ -1,25 +1,26 @@
 #ifndef PROFILER_H
 #define PROFILER_H
 
-extern u64 osClockRate;
+extern u32 osClockRate;
 
 struct ProfilerFrameData {
     /* 0x00 */ s16 numSoundTimes;
     /* 0x02 */ s16 numVblankTimes;
+    /* 0x04 */ s32 padding;          // Add this to keep the 0x08 offset correct
     // gameTimes:
     // 0: thread 5 start
     // 1: level script execution
     // 2: render
     // 3: display lists
     // 4: thread 4 end (0 terminated)
-    /* 0x08 */ OSTime gameTimes[5];
+    /* 0x08 */ u32 gameTimes[5];     // Force to u32
     // gfxTimes:
     // 0: processors queued
     // 1: rsp completed
     // 2: rdp completed
-    /* 0x30 */ OSTime gfxTimes[3];
-    /* 0x48 */ OSTime soundTimes[8];
-    /* 0x88 */ OSTime vblankTimes[8];
+    /* 0x1C */ u32 gfxTimes[3];      // Corrected offset for 32-bit
+    /* 0x28 */ u32 soundTimes[8];    // Corrected offset for 32-bit
+    /* 0x48 */ u32 vblankTimes[8];   // Corrected offset for 32-bit
 };
 
 // thread event IDs
@@ -37,3 +38,4 @@ void resource_display(void);
 extern s32 gEnableResourceMeters;
 
 #endif /* PROFILER_H */
+
